v2 디렉터리에 메시지 단축키 관련 테스트 추가: 클라이언트의 Send 및 List 메서드에 대한 테스트 케이스를 작성하여 메시지 전송 및 목록 조회 기능을 검증함. 기존 테스트에서 쿼리 조합 및 서버 오류 처리에 대한 테스트를 분리하여 코드 구조를 정리함.
